lib: FEATURE: allow setting multiple filters at once
Unify code for processing filter parameter of NETCONF operations.

Filter is now accepted also if it is set as multiple elements on the
same (root) level. This allows setting filter for multiple data models
(devices) at once. Given filter string is not supposed to be well-formed
XML document, but instead one or more XML elements describing the
NETCONF filter.