lib: BUGFIX: rpc-error handling
When a rpc-error reply was received, callback function to process
rpc-error was invoked, but this function is not set by default. So now,
existence of this callback is checked and if the callback is not set,
rpc-error is returned as a nc_reply to be completely processed by caller
application.