lib: CHANGE: multiple configs in one request
We now enable sending multiple elements in copy-config's and
edit-config's <config> parameter. On the server side, this is processed
as separated requests on the specific datastore where root element's
name and namespace match. From the server point of view, this is done
transparently, server only has to call ncds_apply_rpc() to all
datastores and then connect the results into a single reply.