lib: BUGFIX: breaking locks of dummy sessions
libnetconf uses dummy sessions internally for locking/unlocking
datastores. But when such a session is freed, using break_locks lead to
the infinite loop. So now, there is an exception for libnetconf's
internal dummy sessions (with session ID 0) to do not apply break_locks
when such a session is freed. It does not matter, because such session
should not take any datastore's lock.