lib: BUGFIX: using namespace of edit-config's operation
When the namespace of edit-config's operation (or any other namespace
used inside edit-config's <config> element) is defined in higher
elements, xmlNodeDump ignores the definition and we lose it (but we need
it). Therefore, before dumping nodes, we have to recursively copy all
needed nodes (content of the <config>) into the new document - this
operation also copy the namespaces to the right place.

TODO: check all other usage of xmlNodeDump and cover it in the simmilar
way.