lib: BUGFIX: correct xml nodes dumping
When the namespace used in some nested xml element is defined
in a higher element and only a subtree without namespace definition is
dumped using xmlNodeDump(), the definition of the namespace is lost. To
avoid this situation, before dumping nodes, we have to recursively copy
all dumped nodes into the new document - this operation also copy the
namespaces to the right place.

This bug occured when edit-config's operation's namespace is defined
e.g. in the <edit-config> element instead of definition inside the
edit-config's <config> content.