client: BUGFIX: processing copy-config parameters
Change of the NC_DATASTORE enum was not reflected. The enum type now
contains explicit value for config source datastore (when configuration
datastore is set as some xml data), but client still used 0 value -
libnetconf was used to detect it according to specified config
parameter, but now it requires explicit NC_DATASTORE_CONFIG as the
source datastore.