all: FEATURE: add RCSID string
Add RCSID string which specifies the source code and the specific
version (taken from git hash this time). Running ident(1) on the
compiled binaries returns the list of used source codes with the hash of
the last git commit before compilation. If the source codes are compiled
out of the git repository, VERSION value is used.