server: CHANGE: configurable path to server config
Make path to the configuration data model and datastore file used by
server configurable via configure script. So far, the /tmp/ path was
hardcoded and all used files were removed on reboot. Now, the default
path is /usr/share/libnetconf/server/ which contains a short README
about the server expectations to the content of this directory.