lib: BUGFIX: accessing sessions statisitcs shared memory
Commit 07ab154751 changes size of the shared memory segment but the key
of this segment was not changed. Therefore, the new implementation can
try (in some circumstances) to access shared memory segment allocated by
previous version and an error appears. This commit changes the key of
the used shared memory segment so if there is some shared memory segment
allocated by previous version, it will not be used since now (it will be
freed with next reboot) and a new segment with a correct size will be
created..