lib: BUGFIX: correct passing through the session info list
passing through the list of monitored sessions and checking for the
duplicities ignored the last item in the list and this item was not
check for duplicity. Therefore, nc_session_monitor() allowed to add the
same session as a new record if this session was placed in the end of
the list.