lib: API CHANGE: allow RPC creation with future supported capabilities
Modify API of the RPC creation function to fully support ALL basic
capabilities defined in RFC 6241. Any other capability will be suuported
using newly defined function nc_rpc_capability_attr(). This way, the API
of the functions creating RPC can be freezed and only new function or
extension to the NC_CAP_ATTR list will be done in the future in order to
support a new capability.

This commit only introduces API changes in the header file(s). These
changes will be implemented in future commits, so the source codes can
not be compiled for now.