lib: CHANGE: add xpath context to the internal message structures
On any message creation, prepare also a context for the xml structure to
be able simple search in the xml document using xpath queries. Currently
only creation and freeing of this information is implemented, changes to
use xpath instead of traversing pointers in the xml structure will
follow in the future.