lib: CHANGE: add notification namespace into context of all messages
Register Notifications namespace for all received or created messages to
allow proper parsing using xPath and XML document context. The namespace
is registered only once when the message is received or created and all
following processing functions can assume that this is already done and
they can use it in their xPath queries.