lib: BUGFIX: correct NC_EDIT_DEFOP_TYPE values.
If caller do not want to specify optional default-operation parameter of
the edit-config operation, it should use value 0 in
nc_rpc_editconfig()'s default_operation parameter. But the 0 value was
so far assigned to the 'none' (NC_EDIT_DEFOP_NONE) default operation.
Now, the 0 value is assigned to NC_EDIT_DEFOP_NOTSET and value of
NC_EDIT_DEFOP_NONE was changed to 3.