lib: CHANGE: make get_status of internal datastores generic
get_status function for internal datastores was speacial so far and
differed from get_state of standard datastores. Now, it is unified to
allow future simple adding of internal repositories without any special
cases for internal datastore(s). It allso allows to add not only empty
internal datastores, so also internal datastores will held the
configuration data (ACM).

The only speciality is on get-schema operation which must be performed
only once - for the datastore with ID == 0 (the first internal
datastore).