lib: TODO: temporary config data refresh in nacm_start()
This add calling of nacm_config_refresh() into nacm_start(), so
configuration data are refreshed with each nacm_start() (when the RPC is
received) even if the configuration data are not changed.

In future, the changes should be detected automatically and
configuration data will be refreshed only when the datastore changes.