lib: BUGFIX: matching operation/notification names in NACM
In NACM rule specification, operation names (rpc-name) and notification
names (notification-name) have a form of space delimited list of
affected operations/notifications. So far, the matching was checked
using strstr() but it fails for example in case of specified get-config
operation when also get is wrongly matched. The list is now tokenized
into the list of strings with separated names and strcmp() is used for
matching.