lib: FEATURE: add datastore function to detect changes
Datastores are now able to detect if the datastore (e.g. a real file)
has changed since last access to it. Datastore functions as well as a
generic caller are now able to get know if something changed and if they
have to update datastore or some internal structures.

Internal NACM datastore is now also directly available for NACM
subsystem to simplify to accelerate work with NACM configuration data.