lib: BUGFIX: correct implementation of RFC 6536 sec 3.2.4.
If the source of the <copy-config> operation is a datastore, then data
nodes to which the client does not have read access are silently
omitted. We have tried to do some more and do not affect not-readable
data in the target. But now, non-readable data are removed from the
source and complete source datastore is copied into the target. Checking
if the particualr operation with the specific nodes (delete, update,
create) is preserved.