lib: BUGFIX: thread safe reading from different sessions
When client established the first session and subscribed for the
notifications, an another session sometimes (well, in most cases) could
not be established.
REASON: nc_session_read_until() read data into a static buffer (what a
stupid idea) and reading in this function interleaved, so the second
session (being in handshake process) actually saw data from the first
session.

I deserve a beer for this debugging. But honestly, it was my code and
my fault :(