lib: REVERT 60c50254: ncds_apply_rpc2all() with ids_copy parameter
ncds_apply_rpc2all() is marked as not thread safety only temporarily,
so currently working with this function from several threads is NOT a
good idea, because ids[] array can be rewritten, but in future this
will be corrected and each thread will be accessing a separated array.

Temporal problem (with using from multiple threads) is better than
unstable API - when the function will be thread safety, ids_copy
parameter will not be needed any more.