lib: BUGFIX: correct semaphore path transformation
Datastore path used for a semaphore name is transformed by a translation
of '/' to '_'. Problem occures when the path contains multiple instances
of '/' in a row - such a path refers to the same file, but it was
transformed to a different semaphore name. Now this is detected and
multiple '/' are transformed into a single '_'.