build: BUGFIX: pthread detection
Before checking library for semaphores, detect threads support in
general (it can be also used for semaphores). pthread support detection
usually leads to -pthread option in CFLAGS and it is used also in client
Makefile (for threads). If library for semaphores (which is usually
-lpthreads) is detected first, pthread detection believes that nothing
else is not needed, but -lpthreads will be hidden in client Makefile.

This is just a hack, actually example application should have a
separated configures which will be done in future when example
applications will be moved into separated projects.