transapi: BUGFIX: removed duplicit callback calls
xmldiff_add_diff() checks first whether the node that is to be added to diff is not already there. This check is performed only when a newly added argument to this function is non-0. However, xmldiff_recursive() on YIN_TYPE_LIST still makes the recursive calls which are now redundant.