lib: CHANGE: inter-process race-condition when locking in custom ds
Use named semaphore to avoid race condition when doing lock and unlock
operations and we are asking underlying custom datastore plugin for
NETCONF lock info before doing real lock/unlock. This behavior arises
only when the custom datastore do implement is_locked() function.
Otherwise, avoidance of a race condition when multiple processes are
accessing the datastore is up to the custom datastore plugin.

This lock avoid simultaneous lock/unlock access to the all custom
datastore plugin instances, i.e. it locks even if there is no race
condition, because we are not able to detect (considering multiple
processes) which custom datastore is accessed.