transapi: BUGFIX: list should be traversed correctly
Until now it could have happened that nodes with different names compared to the list that was checked were treated as if they had the same name. This name check is executed at the very beginning of the loops so the node name is no longer needed as part of list keys.