lib: FEATURE: flag for announcing data modification from TransAPI callback
TransAPI modules (their callbacks) are now allowed to modify data they
received as function parameters. Sometimes it is needed to change data
requested to be set (password settings in IETF system module). Callback
is now able to signal such a change to libnetconf that then applies
changes into the storage.

The variable in the TransAPI module is called config_modified.

This feature requires TransAPI version 2.