lib: BUGFIX: validation in case of edit-config
Fixes issue 18

When doing validation during edit-config, ignore EXIT_RPC_NOT_APPLICABLE
(which means thhat datastore has no validator) and return success. It
differs from validate, because validate is performed on all datastores
and if there is no validator on any datastore, we report it. In case of
edit-config, all datastores don't need to be affected so we don't know
if there is any datastore with set validators.