CHANGE: file ds: using backud datastore files
When the datastore file is corrupted (invalid XML), libnetconf tries
to use a backup datastore. If no such a backup datastore exists
(datastore is corrupted for the first time), it creates an empty one
(named with mkstemp suffix). Next time, if the main datastore is still
corrupted, first it tries to find a previously used backup datastore.