BUGFIX: with-defaults - removing empty elements
During the with-defaults processing libnetconf creates necessary
elements in XML tree. Sometimes, they are not needed in the end
(with-defaults are not added - e.g. when an instance of a list does
not exists in config data) and libnetconf should remove all created
XML structure. This commit fixes this process, because so far a part
of the created structure (top element in each DFS run) was not removed.

Fixes Issue 24