BUGFIX: session monitoring inconsistency
When libnetconf application crashed and some session was monitored,
it stayed in the shared session monitoring structure forever. Now,
on each get (where the session stats are printed), processes that
added session into the monitoring structure is checked if it is still
alive (according to /proc/). This way the session monitoring structure
is keeped consistent.

This behaviour can be switched on/off using nc_init()'s
NC_INIT_KEEPALIVECHECK flag. If switched off, crashed sessions will
be kept and displayed as active until the libnetconf system close
using nc_close(1).