FEATURE: allow transAPI module to change order of the callbacks execution
So far, transAPI callbacks are executed from leafs to root (callbacks on the
same level are called according to order in paths file). This commit adds
callbacks_order variable into transAPI modules to set this order manually.
By default, callbacks are still executed 'from leafs to root', but now it is
possible to revert this order to 'from root to leafs'.