CHANGE: move callback_order item to another struct
Move callback_order from transapi_data_callbacks structure into a
general transapi structure. This way, the item can be set directly
also for static transAPI modules and it is not duplicated in the
transAPI module source code (once as a standalone variable and once
as part of transapi_data_callbacks structure).