BUGFIX: segfault in special case of transAPI module
When transAPI module itself works with datastores (adds, removes),
it can change the datastore's model (via libnetconf API) by adding
augment models. But there was a problem that we are using list of
YANG keys generated from the previous model and referenced, but
pointing to a freed memory. Now we check this situation and avoid it.