CHANGE: Listening for call home using separated sockets for IPv4/6
Instead of a single IPv6 socket that is able to translate IPv4 addresses
to IPv6, use 2 standalone sockets, one for IPv4 and one for IPv6. It is
needed to establish at least one of them, but in that case only listening
for IPv4 only or IPv6 only will be performed.