BUGFIX: locking libssh2 channel when multiple channels are used
NETCONF sessions using multiple channels from a single ssh session
actually share the mutex for accessing the session. But when closing
one of these sessions the libnetconf freed the mutex as it wasn't
shared. Therefore, there were problems with using such mutex in the
other (still active) NETCONF session that could lead to segfault or
stucking the application on EBUSY.