CHANGE: make verbose_level volatile
In multi-threaded apps, verbose_level can be changed unexpectedly
from different threads, so it should be volatile. Also the type of
the variable was changed from int ot uint8_t (unsigned char), to
make it as small as possible and to make sure that its change will
be atomic (ok, on x86 or x86_64 it should be even with int, but
this increase our chances also on other platforms).