BUGFIX: get-schema default format
RFC 6022 specifies YANG as a default format if not explicitly specified
in <get-schema> request. However, libnetconf uses YIN format natively and
YANG format can be turned off using --disable-yang-schemas - so far, YIN
was used as default format for get-schema. Due to intercompatibility, we
have to change it to YANG even if YANG support was switched off - clients
expects (according to RFC specification), that if format is not specified,
they will get YANG schema.