BUGFIX: lost notifications from the subscription process
When the subscription thread (ncntf_dispatch_send()) was running
in the process that generated some event, this event was hidden for
the subscription thread and it was not sent to the client. It was
because the pointer in the events file was moved with writing to the
end of the file and it was not moved back.

Thanks Gary Highberger for describing the bug.