doc: CHANGE: Switch replaced by if-else construction in transapi tutorial to make it more general.
The top level container is the only configurable element in toaster data
model. This allow us to use direct compartion (switch) as only
XMLDIFF_ADD or XMLDIFF_REM can be set in op parameter of callback for
this container.
But in general container can receive various combinations of event flags
so it is better to check it in 'if' using bitwise and (&).