BUGFIX: removing default="true" elements from config data
1. on rollback after a transAPI callback failure, the config data
could be contamined with default nodes that actually were not in
previous config data (they are just added to data passed to transAPI
callbacks). This patch removes these nodes from the stored
configuration data.

2. RFC 6243 (with-defaults capability) sec. 4.5.2 defines server
behaviour when it supports report-all-tagged mode and configuration
data in edit-config operation contain nodes with default="true"
attribute. Such a node should cause reverting item value to the default
value and then removing this node from the configuration data stored
in the datastore. So far, this was ignored (wrongly implemented).