BUGFIX: edit-config with insert attribute
When we are inserting (or moving) a list item to the position "first"
we have to check if we are inside another list and if so, the inserted
(moved) item cannot be placed as first child since there must be keys.
So in this case, another (first) instance of the inserted (moved) list
item is found and the inserted (moved) item is placed before it.