BUGFIX: Yang MUST statement validation
Schematron generates error report under the two elements:
successful-report and failed-assert. So for, only the first of them
was covered. Now, the failed-assert nodes are detected and reported
as validation failure. The patch also includes a change to attach
all validation errors (from the first wrong datastore) to the
rpc-reply.

Fixes issue 42