BUGFIX: selecting nodes for processing edit-config operation
When edit-config/copy-config request is processed, the configuration
data root elements for the processed datastore part must be selected
according to the matching namespace and the element name. By error,
when the data element had no namespace, it matched with any data model
namespace.

Caused segfault when the with-defaults values were processed.

Fixes issue 43