CHANGE: make lncdatastore build (make) separated
Remove building lncdatastore from the main Makefile. lncdatastore
is an optional development tool not needed for a common use of the
libnetconf library.

This removes a hiden dependency on libreadline which is not needed
for the library itself. The libreadline is required by lncdatastore
and it is written in its README file.

Fixes issue 44