BUGFIX: filling error structure referenced by caller as NULL pointer
Datastore functions fill error structures when an error occures. But
sometimes the returned error was not required so the calling functions
put a NULL pointer and filling the structure in the datastore functions
led to a segfault.

All callers now should put a valid pointer and datastore functions use
assertion to the error parameter.