BUGFIX: reading data from datastore
Data returned as content of the datastore or as status data sometimes
can contain also the XML head (<?xml). In that case we have to ignore
this part when we want to work with it as with an XML subtree. So far,
this check was only in one specific part, but this commit moves this
code into the routine that parses the data and which is used on
multiple places in the code.