CHANGE: optimize creation of RPCs/RPC-REPLYs
Remove a function that actually that actually does nothing. It was
trying to detect various information about the message, but the routines
that were calling this function already knew those parameters. So, now,
the parameters are directly set by the higher level functions and the
function covering that automatic detection was removed.

The automatic detection process is still used when the RPC/RPC-REPLY is
received via a NETCONF session.