BUGFIX: wrong libs paths in lnctool when building RPM
lnctool includes absolut paths to some rng/schematron "libs" being
instolled with it. The paths are expanded on `make install` and it
contains DESTDIR value since it is where the libs are also installed.
However this does not works when we build RPM - DESTDIR is used to
locate files that are put to the RPM package but in the target system
they are installed to the path without the DESTDIR.

We add --with-rpc option to the configure which is used in spec file
(it is hidden from the configure's help). The option causes that paths
in the lnctool are expanded without the DESTDIR included.