NACM: allow NACM turning off
Change nc_init's flags to be specific only for the NACM and do not include
dependencies. This way the flag can be also used to tune off via neg value
without turning off its dependencies.

There is also an exception in import. When the NACM is turned off, the modules
importing NACM are allowed (i.e. the processing do not stop) only with a
warning that the NACM module is turned off. The NACM's extensions are ignored.