BUGFIX: starvation of the notification thread
re-locking of the session in the receive function is too quick and
it can cause starvation of some other thread waiting for the session's
channel (e.g. notification thread).

Since the sending the message is less often than waiting to receive the
message (which is usually part of the main loop), the sending thread now,
besides locking the mutex, set tha flag. Receiving thread (function) check
this flag and if it is set, it instead of polling the session channel waits
for the same timeout as used for polling and then returns WOULDBLOCK (so the
same result as if the poll timeouts).

The result is that the sending thread has a higher priority than the receiving
one.