BUGFIX: generating netconf-config-change notification
Move generating this notification from ncds_apply_rpc() into
ncds_apply_rpc2all(). SO far the notification was generated in case
of modifying the first internal datastore (to avoid generating multiple
notifications for a single rpc). But in many cases this datastore part
is not affected at all and therefore the notification was not generated
at all in the most cases.