CHANGE: Check transAPI version for static modules
So far, libnetconf did not check the version of static transAPI modules.
We need this check since libnetconf can change in the same way as for
dynamic modules. To achieve this, the transapi structure had to be changed,
the version memeber was added at offset 0 to make get know if the memeber
is present or not (previous transAPI versions). If not, the check actually
get some high number since the offset 0 was used to store address of the
modules init function.

This makes API incompatible with 0.9.x.