CHANGE: do not autodetect SSH publickeys
When no public key pair was set to be used, we tried to get the path
automatically according to what is used in OpenSSH. Now, we are moving
from the OpenSSH (Netopeer integrates SSH server using libssh) so it will
be better to do not follow what is used in OpenSSH. It can be confusing when
user set publickey for OpenSSH and when starting Netopeer, it will not work
since the netopeer-server has its own configuration (different from sshd) and
the publickey must be set separately. Therefore it is better to do the same
on the client side.