BUGFIX ext_model consolidation according to features
So far, features that are checked during the consolidation were taken
only from the base module. But in case of augmentation, we need to
apply also features defined in the augmenting data model. By this commit,
this issue is fixed.

fixes #71

But there is another problem - if-feature statement can refer a feature
definition from the imported module (using prefix). This issue is not
covered and such features are not applied. With the current approach, this
issue can be hardly fixed. Future usage of libyang should solve this
problem.