Revert "[artf127329] rotating events files stream
General idea is to rename stream file when exceeds desired size,
and mark it with magic value at the end, so agent which read this
file would know that need to switch to another file, when new file
again exceed size remove first one, rename current, and create
new and so on."

This reverts commit 5ee217b1775a0ab18b0e9303b91039508d340be7.

The implementation has several issues described in #156. For now,
it is removed from master branch and kept for further development
in separate rotating_events branch.