CHANGE add --disable-ncntf-replay configure option
Allow to turn off the replay feature of NETCONF event notifications.
When the replay is disabled, libnetconf handles all streams as if
they have turned off the replay property. However, when a stream is
being created by ncntf_stream_new() with replay parameter 1, it is
correctly stored into the stream file and in case the files is open
by libnetconf with enabled replay option, the stream is able to
do replay. However, libnetconf with turend off the replay feature
does not store the events into any stream file.

Fixes #191