Remove some "if" branches that are always true or always false.
cppcheck noticed that, in src/datastore/edit_config.c, the functions
get_operation() and get_operation_elements() each have an unnecessary
child "if" test whose outcome is known because of a condition in
its parent "if" test.  This change deletes the unnecessary child
tests and a line that is never executed, in the case of the test
that is known to be false always.